backend/pyproject.toml
[project]
name = "money-flow"
version = "0.1.0"
description = ""
authors = [{ name = "teenxsky", email = "kutorgin2002@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.2,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "faker (>=37.3.0,<38.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
ruff = "^0.11.13"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "."
venv = ".venv"
reportAssignmentType = false
reportArgumentType = false
reportOptionalMemberAccess = false
reportAttributeAccessIssue = false
reportIncompatibleVariableOverride = false
reportIndexIssue = false
